requirements.txt
flask>=2.0.1
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.2.0
numpy>=1.21.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
